requirements.txt
fastmcp
openai
python-dotenv
requests
sseclient-py
